AI Core Access
Changed AI Core so it is now only accessible to the Captain, RD, and CE levels of access. CE gets access because Tcomms is located through the AI core. RD and Cap get access for obvious reasons. For some reason it was set so all heads had access, and I think that a scary prospect is a HOS with access to AI.