Soverign
Upgraded defenses for the AI Core on the Soverign class ships. This is due to a normally easily penetrated space on most ships. The core has been reinforced using Rwalls and AI slippers, and now has redundant power connections under each door. It also adds an atmospheric system to the AI core.